Move implementation of EconItemView class to SDK/EconItemView.h
